requirements.txt
streamlit
requests
googlemaps
SpeechRecognition
langchain-google-genai
langchain-core
pydantic
streamlit-js-eval
streamlit-mic-recorder